Insertion sorting
Insertion sorting is that inserts the key value into the appropriate index position.
First, set the key value, compare it with the previous index value and insert it in an appropriate position. At this time, the index is pushed all by one space.
The time complexity is O(n) = n^2, and inefficient alignment algorithm.