Quick sorting
QuickSort is one of the very fast algorithms.
Divide and conquer the array based on an arbitrary pivot. Then iterate recursively.
On average, the time complexity is O(n) = nlogn, but the worst case (if it is already all sorted) is n^2.
On average, it is a fast and memory-efficient sorting algorithm, but in the worst case, it can degrade performance.