Hashtable
Hashtables are efficient data structures for finding and storing data quickly.
You can find the desired value through the key, feels like a dictionary or a phone book.
Hashtables can fast search and adjust dynamic size and store a variety of variables.
The downside is a collision. Different keys could have the same hash value. search time is incresed and additional memory space is needed because it have to store multiple data in that index.
The solution to this problem is the chaining method, which uses a linked list to connect conflicting data to each index.
Additionally, the performance of the Hashtable depends greatly on the quality of the Hashfunction.
* A hashfunction is that changes a key value to an index value.